requirements.txt
google-generativeai==0.8.4
PyPDF2==3.0.1
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.4.0
SpeechRecognition==3.10.4
gTTS==2.5.3
google-generativeai==0.8.4
PyPDF2==3.0.1
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.4.0
SpeechRecognition==3.10.4
gTTS==2.5.3
pygame==2.6.0
PyAudio==0.2.13
Flask==3.0.0
python-dotenv==1.0.1
pdf2image==1.16.3
pymongo==4.10.1
flask-cors==5.0.0